requirements.txt
streamlit==1.35.0
opencv-python-headless==4.10.0.84
mediapipe==0.10.14
numpy==1.26.4
pandas==2.2.1
scikit-learn==1.3.2
plotly==5.18.0
scipy==1.11.4
openai==1.30.1
python-dotenv==1.0.1